% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{breastCompRes}
\alias{breastCompRes}
\title{rankDrugsGwc output data frame}
\format{data frame}
\usage{
data(breastCompRes)
}
\description{
A dataframe returned from the rankDrugsGwc.R function, used in the
documentation examples.
}
\keyword{datasets}
