% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Function to clean gene data}
\usage{
cleanData(geneIds, geneEsts, pvals, byP = TRUE, forRankAndPlot = FALSE)
}
\arguments{
\item{geneIds}{a character vector containing the gene symbols, ensemble IDs, or entrez IDs for the genes to be analyzed}

\item{geneEsts}{a numeric vector containing estimates for the difference in expression between the two phenotyes. Note that positive values of this estimate (tstat, logFC, etc) must correspond to higher gene expression in the phenotype one would like to reverse}

\item{pvals}{P values from a t-test assessing the diferential expression of the genes between the two phenotypes}

\item{byP}{a boolean value specifying whether genes will be kept according to smallest P values (default) or highest abs(estimate)}

\item{forRankAndPlot}{a boolean value specifying whether the data should be cleaned in a manner that will guarantee it works in the rankDrugsGwc function and volcano plot functions (removes genes that can't be mapped to ensemble IDs)}
}
\value{
a data.frame containing unique gene IDs and their corresponding estimates and p values
}
\description{
This function removes genes that have missing geneIDs, geneEsts or pvals, returns the gene symbols, entrez IDs and ensemble IDs for the geneIDs supplied, and removes genes with the same ID by keeping the genes with lower P values (default) or higher abs(estimates)
}
\examples{
data("geneDataGwc")
##keep genes with the smaller p value of genes with the same id
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value)
##keep genes with larger abs(logFC) of genes with the same id. Could also supply entrez ids
geneDataClean = cleanData(geneIds = geneDataGwc$ensembl_id, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value, byP = FALSE)
}
