% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cmapTrialInfo}
\alias{cmapTrialInfo}
\title{CMAP drug clinical trial info}
\format{data frame}
\usage{
data(cmapTrialInfo)
}
\description{
A dataframe containing the clinical trial results for many of the CMAP drugs,
used for benchmarking various drug repurposing techniques and to provide additional
info about the drugs.
}
\keyword{datasets}
