% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTest.R
\name{combineTest}
\alias{combineTest}
\title{File to import functions combineTest which is needed to run PharmacoGx::gwc() in rankDrugsGwc}
\usage{
combineTest(p, weight, method = c("fisher", "z.transform", "logit"),
  hetero = FALSE, na.rm = FALSE)
}
\description{
This should not be necessary and should be dealt with soon.
}
\examples{
add example with CMAP here

}
\keyword{gwcCMap}
