% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGseaDrugData.R
\name{compareGseaDrugData}
\alias{compareGseaDrugData}
\title{Function to determine pathways that are the most significantly enriched in both a data set and in a drug present in a pharmaco set}
\usage{
compareGseaDrugData(pathways, geneStats, geneIds, drugName, drugPert,
  useTstat = TRUE, nperm = 50000)
}
\arguments{
\item{pathways}{list of gene sets to check. Gene IDs in the gene sets must be entrez IDs or ensemble IDs}

\item{geneStats}{the statistics for the genes with names the same as the ids in the supplied pathways or entrez ids if "all", "bp", "cc", "mf" were used in pathways}

\item{drugName}{the name of the drug in the pharmacoSet to perform the analysis on}

\item{drugPert}{a drug perturbation signature (object of class PharmacoSig) with rownames that correspond to either entrez IDs or ensemble IDs for the genes.}

\item{useTstat}{a boolean specifying whether to use the drugs t-stat in fgsea (default is TRUE). If FALSE the drug's estimate for each gene will be used}

\item{nperm}{number of permutations to do, default is 50000}
}
\value{
A table with GSEA results for the drug and data ordered from most significant pathways in the data and the drug to least. Each row corresponds to a tested pathway. See fgsea for info pertaining to the columns
}
\description{
This function computes gsea results for both a data set and a drug. The function returns a table of the results for the pathways that is ordered according to the most significant pathways in both the data and the drug.
Significance is determined by ranking the pathways p adjusted for the drug and the data (lower p adjusted --> lower rank) and then averaging the rank of the pathway in both the drug and the data. Lower average ranks appear higher in the data frame returned from this function.
}
\examples{
data("drugPertEx")
data("geneDataGwc")
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$t, pvals = geneDataGwc$P.Value)
geneStatsDat = geneDataClean$geneEsts
pathOverlapTab = compareGseaDrugData(pathways = "cc", geneStats = geneStatsDat, geneIds = geneDataClean$entrez, drugName = "BRD-K78431006", drugPert = drugPertEx)
}
