% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corWeighted.R
\name{corWeighted}
\alias{corWeighted}
\title{File to import functions corWeighted which is needed to run PharmacoGx::gwc() in rankDrugsGwc}
\usage{
corWeighted(x, y, w, method = c("pearson", "spearman"),
  alternative = c("two.sided", "greater", "less"), nperm = 0, nthread = 1,
  setseed, na.rm = FALSE)
}
\description{
This should not be necessary and should be dealt with soon.
}
\examples{
add example with CMAP here

}
\keyword{gwcCMap}
