% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drugPertEx}
\alias{drugPertEx}
\title{Drug perturbation signature example}
\format{PharmacoSig object}
\usage{
data(drugPertEx)
}
\description{
A small example version of the L1000 PharmacoSet, used in the
documentation examples. There is also a corresponding Pharmacoset object
used to create this signature for use in the examples (psetSub). All credit for the data goes to the
Broad Institute. This is not a full version of the dataset, most of
of the dataset was removed to make runnable example code. For the full dataset,
please download using the downloadPSet function of PharmacoGx.
}
\keyword{datasets}
