% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugRankPlots.R
\name{drugRankPlots}
\alias{drugRankPlots}
\title{Function to compute drug ranks and plot their rank and change in rank over the iterations/as the number of genes in the analysis change}
\usage{
drugRankPlots(drugResults, drugName = NULL)
}
\arguments{
\item{drugResults}{a data frame returned from rankDrugsGwc containing the connectivity scores of the drugs over the iterations}

\item{drugName}{The name of the drug one would like to generate plots for to examine its performance as a function of the number of genes in the analysis. Defaults to the drug with the most negative (most likely to reverse disease phenotype) rank if a name is not provided.}
}
\description{
This function uses the results from rankDrugsGwc to plot the rank of a drug over the iterations and the rank of the change in rank of the drug over the iterations (to see drug stability over the iterations)
}
\examples{
data("geneDataGwc")
data("drugPertEx")
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$t, pvals = geneDataGwc$P.Value)
pFrame = getDrugsPValue(drugNameVec = c("BRD-K78431006", "BRD-A68739437", "BRD-K73368362"), geneIds = geneDataClean$ensemble, geneEsts = geneDataClean$geneEsts, pvals = geneDataClean$pvals, drugPert = drugPertEx, numbIters = 5, numbPerms = 50000)
}
\keyword{gwcCMap}
