% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drugResults}
\alias{drugResults}
\title{rankDrugsGwc output data frame generated from the example}
\format{data frame}
\usage{
data(drugResults)
}
\description{
The dataframe returned from the example in the rankDrugsGwc.R function, generated by passing
drugPertEx and psetSub into the function
}
\keyword{datasets}
