% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{geneDataGwc}
\alias{geneDataGwc}
\title{Differential expression data frame}
\format{data frame}
\usage{
data(geneDataGwc)
}
\description{
a list of differentially expressed genes between a disease state and a normal state, the data has the t-saats,
corresponding p values, and log(fold-change) for each gene. Used to run examples throughout the package.
}
\keyword{datasets}
