% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBenchmarkDrugNames.R
\name{getBenchmarkDrugs}
\alias{getBenchmarkDrugs}
\title{Function to get drugs to use when benchmarking}
\usage{
getBenchmarkDrugs(searchWords, containsAll = FALSE, drugsToKeep = NULL)
}
\arguments{
\item{searchWords}{a character vector containing the key words to be used to select drug names}

\item{containsAll}{a boolean. If set to TRUE, a drug name is only returned if its clinical trial info contains all the search phrases/words. Default is FALSE}

\item{drugsToKeep}{a character vector containing the names of the drugs that should be searched by using the words supplied in searchWords. Used to prevent the returned drug names from containing drugs that wont be used in the benchmarking analysis as they are not in the drug perturbation signature used (ex send in CMap drug names)}
}
\value{
a frame with clinical trial info for the drug names supplied, if available. If no info is available NA is returned
}
\description{
This function will return a frame with drug clinical trial information for drugs that contained the search words inputted to this function in their study information. The function
is most useful for determining drugs to use for benchmarking in the benchMarkingCmap function. For instance, if one tests various
drug repurposing methods on a breast cancer signature, the input search word could be set to "breast" to find breast cancer drugs
to use when benchmarking
}
\examples{
searchWords = c("solid tumor", "depressive")
drugNames = getBenchmarkDrugNames(searchWords)
#below line will show other studies the drugs have been involved in, studies that did not contain the search words
clinicalInfoFrame = getDrugClinicalInfo(drugNames)
}
