% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugClinicalInfo.R
\name{getDrugClinicalInfo}
\alias{getDrugClinicalInfo}
\title{Function to get clinical trial info for drugs}
\usage{
getDrugClinicalInfo(drugNames)
}
\arguments{
\item{drugNames}{a character vector containing the names of the drugs to return clinical trial info for}
}
\value{
a frame with clinical trial info for the drug names supplied, if available. If no info is available NA is returned
}
\description{
This function will return a data frame with clinical trial info for the drug names supplied, if available. Note that importing the data
file "cmapTrialInfo" will give one access to all of the available clinical trial info
}
\examples{
drugNames = c("doxorubicin", "cetuximab")
clinicalInfo = getDrugClinicalInfo(drugNames)
}
