% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugScoresForGene.R
\name{getDrugScoresForGenes}
\alias{getDrugScoresForGenes}
\title{Function to use specified genes in a data set to get connectivity scores and p values for drugs via a genome wide correlation}
\usage{
getDrugScoresForGenes(geneIds, geneEsts, pvals, drugPert = NULL,
  pharmSet = NULL, drugScoreMeth = "gwc", gwcMethod = "spearman",
  numbPerms = 10000, drugEst = TRUE)
}
\arguments{
\item{geneIds}{a character vector containing the gene symbols, ensemble IDs, or entrez IDs for the genes to be analyzed}

\item{geneEsts}{a numeric vector containing estimates for the difference in expression between the two phenotyes. Note that positive values of this estimate (tstat, logFC, etc) must correspond to higher gene expression in the phenotype one would like to reverse}

\item{pvals}{P values from a t-test assessing the diferential expression of the genes between the two phenotypes}

\item{drugPert}{a drug perturbation signature (object of class PharmacoSig) with rownames that correspond to the ensemble IDs of the genes}

\item{pharmSet}{the PharmacoSet used to generate the drug perturbation signature. Supplying this will add additonal info about the drugs in the results table.}

\item{drugScoreMeth}{a string specifying which drug repurposing technique to use to score the drugs during each trial. The options for this parameter are currently "gwc" and "fgsea". Default is "gwc"}

\item{gwcMethod}{a character string specifying which method to use when computing correlations in the gwc function. The options are spearman (default) or pearson.}

\item{numbPerms}{The number of permutations to be used to compute the p value in the gwc function.}

\item{drugEst}{a boolean specifying whether to use the estimates for each gene of the drug perturbation signature in the gwc calculation (TRUE) or to use the t-stats for each gene in the drug perturbation signatur ein the gwc calculation (FALSE). Default is TRUE}

\item{drugNameVec}{a character vector containing the names of the drugs in the drug perturbation signature to compute a more accurate p value for}
}
\value{
a data frame containing scores and p values for the drugs when analyzed using the supplied genes
}
\description{
This function will use the genes inputed to determine connectivity scores and p values for the drugs in a drug perturbation signature
}
\examples{
data("psetSub")
data("geneDataGwc")
data("drugPertEx")
geneDataClean = cleanData(geneIds = geneDataGwc$ensembl_id, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value)
#keep genes with p value < 0.10 and |logFC| > 0.5
rowsToKeep = removeGenes(idsOne = geneDataClean$ensemble, idsTwo = geneDataClean$ensemble, valuesTwo = cbind(geneDataClean$pvals, abs(geneDataClean$geneEsts)), conditionVals = c(0.10, .5), conditionsDirec = c(FALSE, TRUE))
sigGeneData = geneDataClean[rowsToKeep, ]
gwcResults = getDrugScoresForGenes(geneIds = sigGeneData$ensemble, geneEsts = sigGeneData$geneEsts, pvals = sigGeneData$pvals, drugPert = drugPertEx, pharmSet = psetSub)

}
\keyword{gwcCMap}
