% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugsImpactOnGenes.R
\name{getDrugsImpactOnGenes}
\alias{getDrugsImpactOnGenes}
\title{Function to generate heatmaps that show a drugs influence on a set of genes}
\usage{
getDrugsImpactOnGenes(drugPert, pharmSet, mDataType, drugName, genesOfInt,
  makePlot = TRUE)
}
\arguments{
\item{drugPert}{the drug perturbation signature of interest}

\item{pharmSet}{the pharmacoSet used to generate the drug perturbation signature}

\item{mDataType}{the type of molecular data to obtain from the PharmacoSet (see PharmacoGx::molecularProfiles)}

\item{drugName}{the name of the drug one would like to inspect for its effect on the supplied genes}

\item{genesOfInt}{a character vector containing the names of the genes one would like to observe the drugs effect on}

\item{makePlot}{a boolean specifying whether to generate the plot showing the gene expression changes due to the drug. Default is TRUE}
}
\value{
a list containing 1. beforeDrugMat which is a matrix containing the expression levels for the genes before the drug was applied.
 2. afterDrugMat which is a matrix containing the expression levels for the genes before the drug was applied. 3. geneChangePlot, which is a
 levelplot object showing the change in expression for the genes due to the drugs
}
\description{
This function returns a heat map that shows how a drug changes the expression of a set of genes, as determined by molecular profiles in the pharmacoSet supplied.
}
\examples{
data("drugPertEx")
data("psetSub")
#example in getSigGeneVolPlot function yields below genes
drivingGenes = c("ABCC5", "VPS28", "RPS6KA2", "TRIM2", "CD47", "MELK", "CDKN3", "GGH", "MCM7", "KIF2C")
plotAndMats = getDrugsImpactOnGenes(drugPert = drugPertEx, pharmSet = psetSub, mDataType = "rna", drugName = "BRD-K78431006", genesOfInt = drivingGenes)

}
