% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugsPValue.R
\name{getDrugsPValue}
\alias{getDrugsPValue}
\title{Function to compute p values for the drugs}
\usage{
getDrugsPValue(drugNameVec, geneIds, geneEsts, pvals, drugPert = NULL,
  drugScoreMeth = "gwc", pCut = TRUE, cutOff = 0.05, genesToStart = 0.2,
  numbIters = 10, gwcMethod = "spearman", numbPerms = 1e+05,
  drugEst = TRUE, extraData = NULL, extraCut = NULL, extraDirec = NULL)
}
\arguments{
\item{drugNameVec}{a character vector containing the names of the drugs in the drug perturbation signature to compute a more accurate p value for}

\item{geneIds}{a character vector containing the gene symbols, ensemble IDs, or entrez IDs for the genes to be analyzed}

\item{geneEsts}{a numeric vector containing estimates for the difference in expression between the two phenotyes. Note that positive values of this estimate (tstat, logFC, etc) must correspond to higher gene expression in the phenotype one would like to reverse}

\item{pvals}{P values from a t-test assessing the diferential expression of the genes between the two phenotypes}

\item{drugPert}{a drug perturbation signature (object of class PharmacoSig) with rownames that correspond to the ensemble IDs of the genes}

\item{drugScoreMeth}{a string specifying which drug repurposing technique to use to score the drugs during each trial. The options for this parameter are currently "gwc" and "fgsea". Default is "gwc"}

\item{pCut}{a boolean specifying whether to use pvalues to remove insignificant genes from the CMAP analysis (TRUE) or to remove genes from the CMAP analysis according to their supplied gene estimates (FALSE)}

\item{cutOff}{if pCut is TRUE then this value represents the p value threshold used to filter out genes. If pCut is FALSE then this value represents the fraction of genes present in both the data and drug perturbation signature with the top absolute value of gene estimates that will be left in the analysis. cutOff should be between 0 and 1.}

\item{genesToStart}{a value between 0 and 1 representing the fraction of significant genes present in the data and drug perturbation signature to use in the first iteration of the CMAP analysis. Recommended to be at least 0.15 to avoid large changes during the early iterations due to having too few genes present.}

\item{numbIters}{The number of iterations that will occur in the analysis. numbIters will set the rate at which genes are added to the analysis.}

\item{gwcMethod}{a character string specifying which method to use when computing correlations in the gwc function. The options are spearman (default) or pearson.}

\item{numbPerms}{The number of permutations to be used to compute the p value in the gwc function.}

\item{drugEst}{a boolean specifying whether to use the estimates for each gene of the drug perturbation signature in the gwc calculation (TRUE) or to use the t-stats for each gene in the drug perturbation signatur ein the gwc calculation (FALSE). Default is TRUE}

\item{extraData}{a data.frame where each column represents values one would like to inspect to determine if the genes corresponding to these values should be removed if they meet the conditions specified in extraCut and extraDirec. Useful if one would like to remove genes based on logFC or other values. extra data must have the same number of rows as the length of vectors geneIds, geneEsts, and pvals.}

\item{extraCut}{a numeric vector, with the nth value corresponding to the nth column of the m by n data.frame supplied in extraData, where each value indicates the value that needs to be reached for the columns of the data fram in order for the ids with that value to be kept or removed. Whether removal occurs when a value is greater or less than the value in this vector is specified in the extraDirec variable.}

\item{extraDirec}{a boolean vector, where TRUE means that ids whose value is greater than that specified in extraVals will be removed and FALSE means those with a value less than conditionVals will be removed.}
}
\value{
a data frame containing the p value from the connectivity score computed using the number of genes specified in the rows for each drug.
}
\description{
This function will compute the p values for only the drugs inputted via drugNameVec. To save time, users are recommended to run rankDrugsGwc with a low nperm and then use this function with a high nperm (~100,000) to get the p values for the top drug candidates.
}
\examples{
data("geneDataGwc")
data("drugPertEx")
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$t, pvals = geneDataGwc$P.Value)
pFrame = getDrugsPValue(drugNameVec = c("BRD-K78431006", "BRD-A68739437", "BRD-K73368362"), geneIds = geneDataClean$ensemble, geneEsts = geneDataClean$geneEsts, pvals = geneDataClean$pvals, drugPert = drugPertEx, numbIters = 5, numbPerms = 50000)
}
\keyword{gwcCMap}
