% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigGeneVolcPlot.R
\name{getSigGeneVolcPlot}
\alias{getSigGeneVolcPlot}
\title{Function to determine which genes drove a drugs connectivity score and create volcano plots that shows which genes drove a drugs connectivity score}
\usage{
getSigGeneVolcPlot(genesEst, genesSymb, genesP, genesId, drugPert, drugName,
  drugScoreMeth, gwcMethod, droveNegScore, drugEst, drugVolcPlot = FALSE,
  supressPlot = TRUE)
}
\arguments{
\item{genesEst}{a numeric vector with the estimates (typically LogFC) for the genes}

\item{genesSymb}{a character vector with the gene symbols for the gene estimates provided}

\item{genesP}{a numeric vector with the p values for the gene estimates provided}

\item{genesId}{a character vector with gene Ids for the gene estimates provided. Must be the same type of ID as those used for the row names of the drugPert variable}

\item{drugPert}{the drug perturbation signature used in the analysis}

\item{drugName}{the name of the drug in the drug perturbation signature that one would like to inspect}

\item{drugScoreMeth}{a string specifying which drug repurposing technique was used to score the drugs during each trial. The options for this parameter are currently "gwc" and "fgsea"}

\item{gwcMethod}{a character string specifying which method was used when computing correlations in the gwc function. The options are spearman or pearson.}

\item{droveNegScore}{a boolean specifying whether the plots should label the genes that drove the negative connectivity score (reversed the phenotype) between the drug and the data. If FALSE, the genes that drove the positive connectivity score will be labelled.}

\item{drugEst}{a boolean specifying whether the estimates for each gene of the drug perturbation signature were used in the gwc calculation (TRUE) or if the t-stats for each gene in the drug perturbation signature were used in the gwc calculation (FALSE).}

\item{drugVolcPlot}{a boolean (default is FALSE) specifying whether to generate a volcano plot for the top and bottom 5 genes (by estimate) in the drug that were present in the data supplied}

\item{supressPlot}{a boolean specifying whether to generate volcano plots that illustrate the genes driving the connectivity score. TRUE by default}
}
\value{
the names of the top 5 and bottom 5 genes that drove the connectivity score (top 5 first, then bottom 5)
}
\description{
This function displays volcano plots for the data and drug supplied with the top 5 and bottom 5 genes that drove the connectivity score labelled. Top 5 and bottom 5 here refer to genes that had positive and negative gene estimates in the data, respectively, and drove the connectivity score the most.
}
\examples{
data("geneDataGwc")
data("drugPertEx")
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value, forRankAndPlot = TRUE)
drivingGenes = getSigGeneVolcPlot(genesEst = geneDataClean$geneEsts, genesSymb = geneDataClean$symbol, genesP = geneDataClean$pvals, genesId = geneDataClean$ensemble, drugPert = drugPertEx, drugName = "BRD-K78431006", gwcMethod = "pearson", droveNegScore = TRUE, drugEst = TRUE, drugVolcPlot = FALSE, supressPlot = FALSE)

}
\keyword{gwcCMap}
