% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueIds.R
\name{getUniqueIds}
\alias{getUniqueIds}
\title{Function to determine which genes to keep when there are multiples IDs per gene with different values}
\usage{
getUniqueIds(geneIds, geneVals, byP)
}
\arguments{
\item{geneIds}{a character vector containing the gene symbols, ensemble IDs, or entrez IDs for the genes to be analyzed}

\item{geneVals}{a numeric vector containing values or p values for the genes with the corresponding gene IDs}

\item{byP}{a boolean value specifying whether genes will be kept according to smallest P values or highest abs(value)}
}
\value{
a numeric vector containing the indices for the geneId elements that should be kept
}
\description{
This function returns the indices of genes in a vector that should be kept. Genes that should be kept are those without other genes having the same ID and genes with lower P values (default) or higher abs(estimates) of those genes with the same ID.
}
\examples{
data("geneDataGwc")
indsToKeep = getUniqueIds(geneIds = as.character(geneDataGwc$ensembl_id), geneVals = geneDataGwc$P.Value, byP = TRUE)
##or keep genes with larger abs(logFC)
indsToKeep = getUniqueIds(geneIds = as.character(geneDataGwc$ensembl_id), geneVals = geneDataGwc$logFC, byP = FALSE)
geneDataUnique = geneDataGwc[indsToKeep, ]
}
