% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDrugResults.R
\name{inspectDrugResults}
\alias{inspectDrugResults}
\title{Function to inspect drug candidates}
\usage{
inspectDrugResults(genesEst, genesSymb, genesP, genesId, drugPert, pharmSet,
  drugScoreMeth, mDataType, drugResults, gwcMethod, drugEst,
  drugVolcPlot = FALSE, topDrugName = NULL, botDrugName = NULL,
  showMimic = FALSE)
}
\arguments{
\item{genesEst}{a numeric vector with the estimates (typically LogFC) for the genes}

\item{genesSymb}{a character vector with the gene symbols for the gene estimates provided}

\item{genesP}{a numeric vector with the p values for the gene estimates provided}

\item{genesId}{a character vector with gene Ids for the gene estimates provided}

\item{drugPert}{the drug perturbation signature used in the analysis}

\item{pharmSet}{the pharmacoSet used to generate the drug perturbation signature}

\item{drugScoreMeth}{a string specifying which drug repurposing technique was used to score the drugs during each trial. The options for this parameter are currently "gwc" and "fgsea"}

\item{mDataType}{the type of molecular data to obtain from the PharmacoSet (see PharmacoGx::molecularProfiles)}

\item{drugResults}{a drug rank data frame (such as that returned by rankDrugsGwc) ordered by most negative to most positive scored drugs with the rownames containing the drug names}

\item{gwcMethod}{a character string specifying which method was used when computing correlations in the gwc function. The options are spearman or pearson.}

\item{drugEst}{a boolean specifying whether the estimates for each gene of the drug perturbation signature were used in the gwc calculation (TRUE) or if the t-stats for each gene in the drug perturbation signature were used in the gwc calculation (FALSE).}

\item{drugVolcPlot}{a boolean (default is FALSE) specifying whether to generate a volcano plot for the top and bottom 5 genes (by estimate) in the drug that were present in the data supplied}

\item{topDrugName}{the name of the drug that one would like to check reverses the disease phenotype. If not provided, the best candidate according to the drugResults frame will be selected.}

\item{botDrugName}{the name of the drug that one would like to check enhances the disease phenotype. If not provided, the best candidate according to the drugResults frame will be selected.}

\item{showMimic}{a boolean (default is FALSE) specifying whether to show plots for the drug that upregulates and down regulates the genes that are overexpressed and under expressed in the disease state.}
}
\description{
This function generates plots that will allow one to verify that the drug candidates that should reverse and enhance the disease phenotype are in fact doing so
}
\examples{
data("geneDataGwc")
data("drugPertEx")
data("psetSub")
data("drugResults")
geneDataClean = cleanData(geneIds = geneDataGwc$ensembl_id, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value, forRankAndPlot = TRUE)
inspectDrugResults(genesEst = geneDataClean$geneEsts, genesSymb = geneDataClean$symbol, genesP = geneDataClean$pvals, genesId = geneDataClean$ensemble, drugPert = drugPertEx, pharmSet = psetSub, mDataType = "rna", drugResults = drugResults, gwcMethod = "pearson", drugEst = TRUE)
}
