% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVolcPlot.R
\name{makeVolcPlot}
\alias{makeVolcPlot}
\title{Function to Create a Volcano Plot}
\usage{
makeVolcPlot(genesEst, genesSymb, genesP, genesId = NULL, idsOfInt = NULL)
}
\arguments{
\item{genesSymb}{a character vector with the gene symbols for the gene estimates provided}

\item{genesP}{a numeric vector with the p values for the gene estimates provided}

\item{genesId}{a character vector with gene Ids for the gene estimates provided. The gene Ids must be of the same type as IdsOfInt. If not provided genes outside of IdsOfInt cannot be removed}

\item{idsOfInt}{a list of Ids to be kept when making the volcano plots, if not provided all Ids will be used}

\item{geneEst}{a numeric vector with the estimates (typically LogFC) for the genes}
}
\description{
This function will create a volcano plot that has the top and bottom 5 genes (according to gene estimates) labelled
}
\examples{
add example with CMAP here
data("geneDataGwc")
geneDataClean = cleanData(geneIds = geneDataGwc$ensembl_id, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value, forRankAndPlot = TRUE)
makeVolcPlot(genesEst = geneDataClean$geneEsts, genesSymb = geneDataClean$symbol, genesP = geneDataClean$pvals)
##only use genes that are present in the pharmacoSet
makeVolcPlot(genesEst = geneDataClean$geneEsts, genesSymb = geneDataClean$symbol, genesP = geneDataClean$pvals, genesId = geneDataClean$ensemble, idsOfInt = rownames(drugPertEx))

}
\keyword{gwcCMap}
