% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainPset.R
\name{obtainPset}
\alias{obtainPset}
\title{Function to download PSets for analysis (credit to PharmacoGx::downloadPset)}
\usage{
obtainPset(psetName = "CMAP")
}
\arguments{
\item{psetName}{The default is "CMAP" but "L1000_compounds" can also be used to obtain the L1000 PharmacoSig object}
}
\value{
a PSet object
}
\description{
The present package requires perturbation datasets (PharmacoSig objects with drug/gene interaction info) to obtain drug lists. These can be obtained
through this function which calls the pharmacoGx::downloadPSet function. Note that if the corresponding pharmacoSet object that was used to create
the pharmacoSig is also obtained it can be used to obtain additional plots that show the impact of the drugs on the individual genes via the seeDrugImpactGenes function
}
