% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDrugGeneInteraction.R
\name{plotDrugGeneInteraction}
\alias{plotDrugGeneInteraction}
\title{Function to determine which genes drove a drugs connectivity score and create volcano plots that shows which genes drove a drugs connectivity score}
\usage{
plotDrugGeneInteraction(genesEst, genesSymb, genesP, genesId, drugPert,
  drugName, drugScoreMeth, gwcMethod, droveNegScore, drugEst,
  numbTopGenes = 5, makePlot = TRUE, drugVolcPlot = FALSE,
  supressPlot = TRUE)
}
\arguments{
\item{genesEst}{a numeric vector with the estimates (typically LogFC) for the genes}

\item{genesSymb}{a character vector with the gene symbols for the gene estimates provided}

\item{genesP}{a numeric vector with the p values for the gene estimates provided}

\item{genesId}{a character vector with gene Ids for the gene estimates provided. Must be the same type of ID as those used for the row names of the drugPert variable}

\item{drugPert}{the drug perturbation signature used in the analysis}

\item{drugName}{the name of the drug in the drug perturbation signature that one would like to inspect}

\item{drugScoreMeth}{a string specifying which drug repurposing technique was used to score the drugs during each trial. The options for this parameter are currently "gwc" and "fgsea"}

\item{gwcMethod}{a character string specifying which method was used when computing correlations in the gwc function. The options are spearman or pearson.}

\item{droveNegScore}{a boolean specifying whether the plots should label the genes that drove the negative connectivity score (reversed the phenotype) between the drug and the data. If FALSE, the genes that drove the positive connectivity score will be labelled.}

\item{drugEst}{a boolean specifying whether the estimates for each gene of the drug perturbation signature were used in the gwc calculation (TRUE) or if the t-stats for each gene in the drug perturbation signature were used in the gwc calculation (FALSE).}

\item{numbTopGenes}{number of top positive and negative genes to return information for and include in the plots}

\item{makePlot}{a boolean specifying whether or not to display the plots showing the genes that the drug interacts with to drive the score. Default is TRUE}

\item{drugVolcPlot}{a boolean (default is FALSE) specifying whether to generate a volcano plot for the top and bottom 5 genes (by estimate) in the drug that were present in the data supplied. Note that if TRUE the other plots are not generated.}

\item{supressPlot}{a boolean specifying whether to generate volcano plots that illustrate the genes driving the connectivity score. TRUE by default}
}
\value{
The information used to generate the plots. Specifically, 2 lists.  One that contains the experiment/user supplied data, a subset of this data for the top positive (according to genesEst)
genes that drove the score by being altered by the drug, and the same for the most negative genes from the data. The second list is the same but for the drugs data as given by drugPert.
The genes will be the same but the info now pertains to how well the drug reduces the expression of these genes.
}
\description{
This function displays volcano plots for the data and drug supplied with the top N and bottom N genes that drove the connectivity score labelled. Top N and bottom N here refer to genes that had positive and negative gene estimates in the data, respectively, and drove the connectivity score the most.
Note that which genes drive the connectivity score can vary with the method used to calculate the score. However, in general these genes are the top positive and negative genes that the drug lowered and raised the expression of, respectively.
}
\examples{
data("geneDataGwc")
data("drugPertEx")
geneDataClean = mapGenes(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$logFC, pvals = geneDataGwc$P.Value, forRankAndPlot = TRUE)
drugGeneInfo = plotDrugGeneInteraction(genesEst = geneDataClean$geneEsts, genesSymb = geneDataClean$symbol, genesP = geneDataClean$pvals, genesId = geneDataClean$ensemble, drugPert = drugPertEx, drugName = "BRD-K78431006", drugScoreMeth = "gwcCmapBox", gwcMethod = "pearson", droveNegScore = TRUE, drugEst = TRUE, drugVolcPlot = FALSE, supressPlot = FALSE)

}
\keyword{gwcCMap}
