% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psetSub}
\alias{psetSub}
\title{A Small L100 Example PharmacoSet}
\format{PharmacoSet object}
\usage{
data(psetSub)
}
\description{
A small example version of the L1000 PharmacoSet, used in the
documentation examples. There is also a corresponding drug perturbation signature
created from this data for use in the examples (drugPertEx). All credit for the data goes to the
Broad Institute. This is not a full version of the dataset, most of
of the dataset was removed to make runnable example code. For the full dataset,
please download using the downloadPSet function of PharmacoGx.
}
\keyword{datasets}
