% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFgsea.R
\name{runFgsea}
\alias{runFgsea}
\title{Function to compute a gene set enrichment analysis on a data set}
\usage{
runFgsea(pathways, geneStats, geneIds, nperm = 50000)
}
\arguments{
\item{pathways}{list of gene sets to check. If supplying the list ensure the stats variable names are the same as the names within the pathways variable list elements. Enter "all", "bp", "cc", "mf" to have all GO pathways, the biological processes pathways, the cellular component pathways, or the molecular functions pathways used, respectively.}

\item{geneStats}{the statistics for the genes}

\item{geneIds}{the Ids for the genes. Ids must be the same as the ids in the supplied pathways or entrez ids if "all", "bp", "cc", "mf" were used in pathways}

\item{nperm}{number of permutations to do, default is 50000}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway. See fgsea for info pertaining to the columns
}
\description{
This function removes genes from data set one from data set two according to the supplied conditions on the values sent to the function
}
\examples{
data("geneDataGwc")
geneDataClean = cleanData(geneIds = geneDataGwc$symbol, geneEsts = geneDataGwc$t, pvals = geneDataGwc$P.Value)
geneStatsDat = geneDataClean$geneEsts
dataGseaResults = runFgsea(pathways = "cc", geneStats = geneStatsDat, geneIds = geneDataClean$entrez)
}
