% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFgseaPharmDrug.R
\name{runFgseaPharmDrug}
\alias{runFgseaPharmDrug}
\title{Function to compute a gene set enrichment analysis on a drug in a pharmacoSet}
\usage{
runFgseaPharmDrug(pathways, drugName, drugPert, useTstat = TRUE,
  nperm = 50000)
}
\arguments{
\item{pathways}{list of gene sets to check. Gene IDs in the gene sets must be entrez IDs or ensemble IDs}

\item{drugName}{the name of the drug in the pharmacoSet to perform the analysis on}

\item{drugPert}{a drug perturbation signature (object of class PharmacoSig) with rownames that correspond to either entrez IDs or ensemble IDs for the genes.}

\item{useTstat}{a boolean specifying whether to use the drugs t-stat in fgsea (default is TRUE). If FALSE the drug's estimate for each gene will be used}

\item{nperm}{number of permutations to do, default is 50000}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway. See fgsea for info pertaining to the columns
}
\description{
This function removes genes from data set one from data set two according to the supplied conditions on the values sent to the function
}
\examples{
data("drugPertEx")
drugGseaResults = runFgseaPharmDrug(pathways = "cc", drugName = "BRD-K78431006", drugPert = drugPertEx)
}
