% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeDrugImpactGenes.R
\name{seeDrugImpactGenes}
\alias{seeDrugImpactGenes}
\title{Function to generate heatmaps that show a drugs influence on a set of genes}
\usage{
seeDrugImpactGenes(drugPert, pharmSet, mDataType, drugName, genesOfInt)
}
\arguments{
\item{drugPert}{the drug perturbation signature of interest}

\item{pharmSet}{the pharmacoSet used to generate the drug perturbation signature}

\item{mDataType}{the type of molecular data to obtain from the PharmacoSet (see PharmacoGx::molecularProfiles)}

\item{drugName}{the name of the drug one would like to inspect for its effect on the supplied genes}

\item{genesOfInt}{a character vector containing the names of the genes one would like to observe the drugs effect on}
}
\description{
This function returns a heat map that shows how a drug changes the expression of a set of genes, as determined by molecular profiles in the pharmacoSet supplied.
}
\examples{
data("drugPertEx")
data("cmapSmall")
#example in getSigGeneVolPlot function yields below genes
drivingGenes = c("ABCC5", "VPS28", "RPS6KA2", "TRIM2", "CD47", "MELK", "CDKN3", "GGH", "MCM7", "KIF2C")  
seeDrugImpactGenes(drugPert = drugPertEx, pharmSet = cmapSmall, mDataType = "rna", drugName = "Prestwick-675", genesOfInt = drivingGenes)

}
