% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkDrugResults.R
\name{benchmarkDrugResults}
\alias{benchmarkDrugResults}
\title{Function to benchmark drug repurposing schemes}
\usage{
benchmarkDrugResults(compResList)
}
\arguments{
\item{drugNames}{a character vector containing the names of the drugs to return clinical trial info for}
}
\value{
a frame with clinical trial info for the drug names supplied, if available. If no info is available NA is returned
}
\description{
This function will return a data frame with
}
\examples{
drugNames = c("doxorubicin", "cetuximab")
clinicalInfo = getDrugClinicalInfo(drugNames)
}
